+++ CHANGELOG.md
@@ -133,6 +133,7 @@ CHANGELOG
 * @gkatsev added a mouse-hover time display to the progress bar ([view](https://github.com/videojs/video.js/pull/2569))
 * @heff added an attributes argument to createEl() ([view](https://github.com/videojs/video.js/pull/2589))
 * @heff made tech related functions private in the player ([view](https://github.com/videojs/video.js/pull/2590))
+* @heff removed the loadedalldata event ([view](https://github.com/videojs/video.js/pull/2591))
 
 --------------------